refactor(lix): promote transaction/types.rs to a top-level module
`transaction/types.rs` was the crate's shared write-row vocabulary living
inside the write-path module that consumes it. 19 of the 20 inbound
back-references to `transaction` targeted it, which made six modules cyclic
with `transaction` for no structural reason.

- `transaction/types.rs` -> `transaction_types.rs`, a git-recorded rename. The
  file is byte-identical to its predecessor apart from the extraction below,
  two imports that extraction leaves unused, and the
  `crate::transaction::types` -> `crate::transaction_types` path rewrite.
- The `StagedCommitChange*` cluster moves the other way, into
  `transaction/staged_commit_changes.rs`. Those five types were the only part
  of the file that named the write path -- `StagedCommitChangeRefs` owns the
  certified `staging::OrderedMutationJournal` -- and nothing outside
  `transaction` reads a `StagedCommitChangeRefs`. Moving the whole cluster
  keeps every field and method visibility exactly as it was.
- `module_layers.rs`: `branch`, `checkpoint`, `collection_generation`,
  `plugin` and `undo_redo` move out of UNLAYERED_MODULES into MODULE_LAYERS.
  ALLOWED_UPWARD_EDGES stays empty.

Mutual module cycles in production code: 10 before, 6 after.

No logic changes, no behaviour changes.
